refactored module and restructured some code
